Refactor BlobSidecarsByRootRequestMessageSchema.java to be not inline and fix usages
