+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/schemas/SchemaDefinitionsDeneb.java
@@ -48,7 +48,7 @@ import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadHeaderSch
 import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadSchema;
 import tech.pegasys.teku.spec.datastructures.execution.versions.deneb.ExecutionPayloadHeaderSchemaDeneb;
 import tech.pegasys.teku.spec.datastructures.execution.versions.deneb.ExecutionPayloadSchemaDeneb;
-import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.BlobSidecarsByRootRequestMessage;
+import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.BlobSidecarsByRootRequestMessageSchema;
 import tech.pegasys.teku.spec.datastructures.operations.SignedBlsToExecutionChangeSchema;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconStateSchema;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.versions.deneb.BeaconStateDeneb;
@@ -85,8 +85,7 @@ public class SchemaDefinitionsDeneb extends SchemaDefinitionsCapella {
   private final BlobsBundleSchema blobsBundleSchema;
   private final BlindedBlobsBundleSchema blindedBlobsBundleSchema;
   private final ExecutionPayloadAndBlobsBundleSchema executionPayloadAndBlobsBundleSchema;
-  private final BlobSidecarsByRootRequestMessage.BlobSidecarsByRootRequestMessageSchema
-      blobSidecarsByRootRequestMessageSchema;
+  private final BlobSidecarsByRootRequestMessageSchema blobSidecarsByRootRequestMessageSchema;
 
   public SchemaDefinitionsDeneb(final SpecConfigDeneb specConfig) {
     super(specConfig.toVersionDeneb().orElseThrow());
@@ -154,8 +153,7 @@ public class SchemaDefinitionsDeneb extends SchemaDefinitionsCapella {
     this.executionPayloadAndBlobsBundleSchema =
         new ExecutionPayloadAndBlobsBundleSchema(executionPayloadSchemaDeneb, blobsBundleSchema);
     this.blobSidecarsByRootRequestMessageSchema =
-        new BlobSidecarsByRootRequestMessage.BlobSidecarsByRootRequestMessageSchema(
-            specConfig.getNetworkingConfig());
+        new BlobSidecarsByRootRequestMessageSchema(specConfig.getNetworkingConfig());
   }
 
   public static SchemaDefinitionsDeneb required(final SchemaDefinitions schemaDefinitions) {
@@ -296,8 +294,7 @@ public class SchemaDefinitionsDeneb extends SchemaDefinitionsCapella {
     return executionPayloadAndBlobsBundleSchema;
   }
 
-  public BlobSidecarsByRootRequestMessage.BlobSidecarsByRootRequestMessageSchema
-      getBlobSidecarsByRootRequestMessageSchema() {
+  public BlobSidecarsByRootRequestMessageSchema getBlobSidecarsByRootRequestMessageSchema() {
     return blobSidecarsByRootRequestMessageSchema;
   }
