+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/peers/DefaultEth2Peer.java
@@ -54,6 +54,7 @@ import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.BeaconBlocksB
 import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.BlobIdentifier;
 import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.BlobSidecarsByRangeRequestMessage;
 import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.BlobSidecarsByRootRequestMessage;
+import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.BlobSidecarsByRootRequestMessageSchema;
 import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.EmptyMessage;
 import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.GoodbyeMessage;
 import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.PingMessage;
@@ -61,6 +62,7 @@ import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.RpcRequest;
 import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.StatusMessage;
 import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.metadata.MetadataMessage;
 import tech.pegasys.teku.spec.datastructures.state.Checkpoint;
+import tech.pegasys.teku.spec.schemas.SchemaDefinitionsDeneb;
 
 class DefaultEth2Peer extends DelegatingPeer implements Eth2Peer {
   private static final Logger LOG = LogManager.getLogger();
@@ -225,7 +227,10 @@ class DefaultEth2Peer extends DelegatingPeer implements Eth2Peer {
         .map(
             method ->
                 requestStream(
-                    method, new BlobSidecarsByRootRequestMessage(blobIdentifiers), listener))
+                    method,
+                    new BlobSidecarsByRootRequestMessage(
+                        getBlobSidecarsByRootRequestMessageSchema(), blobIdentifiers),
+                    listener))
         .orElse(failWithUnsupportedMethodException("BlobSidecarsByRoot"));
   }
 
@@ -256,6 +261,7 @@ class DefaultEth2Peer extends DelegatingPeer implements Eth2Peer {
                 requestOptionalItem(
                     method,
                     new BlobSidecarsByRootRequestMessage(
+                        getBlobSidecarsByRootRequestMessageSchema(),
                         Collections.singletonList(blobIdentifier))))
         .orElse(failWithUnsupportedMethodException("BlobSidecarsByRoot"));
   }
@@ -414,6 +420,12 @@ class DefaultEth2Peer extends DelegatingPeer implements Eth2Peer {
   private <T> SafeFuture<T> failWithUnsupportedMethodException(final String method) {
     return SafeFuture.failedFuture(
         new UnsupportedOperationException(method + " method is not supported"));
+  }
+
+  private BlobSidecarsByRootRequestMessageSchema getBlobSidecarsByRootRequestMessageSchema() {
+    return SchemaDefinitionsDeneb.required(
+            spec.forMilestone(SpecMilestone.DENEB).getSchemaDefinitions())
+        .getBlobSidecarsByRootRequestMessageSchema();
   }
 
   @Override
