+++ networking/eth2/src/test/java/tech/pegasys/teku/networking/eth2/rpc/beaconchain/methods/BlobSidecarsByRootMessageHandlerTest.java
@@ -48,6 +48,7 @@ import tech.pegasys.teku.spec.config.SpecConfigBellatrix;
 import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.BlobSidecar;
 import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.BlobIdentifier;
 import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.BlobSidecarsByRootRequestMessage;
+import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.BlobSidecarsByRootRequestMessageSchema;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
 import tech.pegasys.teku.storage.client.CombinedChainDataClient;
 
@@ -58,6 +59,8 @@ public class BlobSidecarsByRootMessageHandlerTest {
   private final Spec spec = TestSpecFactory.createMinimalWithDenebForkEpoch(denebForkEpoch);
   private final int maxChunkSize =
       SpecConfigBellatrix.required(spec.getGenesisSpecConfig()).getMaxChunkSizeBellatrix();
+  private final BlobSidecarsByRootRequestMessageSchema schema =
+      new BlobSidecarsByRootRequestMessageSchema(spec.getNetworkingConfig());
   private final RpcEncoding rpcEncoding = RpcEncoding.createSszSnappyEncoding(maxChunkSize);
 
   private final String protocolId =
@@ -117,7 +120,7 @@ public class BlobSidecarsByRootMessageHandlerTest {
     final int maxRequestBlobSidecars = getMaxRequestBlobSidecars();
     final BlobSidecarsByRootRequestMessage request =
         new BlobSidecarsByRootRequestMessage(
-            dataStructureUtil.randomBlobIdentifiers(maxRequestBlobSidecars + 1));
+            schema, dataStructureUtil.randomBlobIdentifiers(maxRequestBlobSidecars + 1));
 
     final Optional<RpcException> result = handler.validateRequest(protocolId, request);
 
@@ -145,7 +148,7 @@ public class BlobSidecarsByRootMessageHandlerTest {
     when(peer.popBlobSidecarRequests(callback, 5)).thenReturn(false);
 
     final BlobSidecarsByRootRequestMessage request =
-        new BlobSidecarsByRootRequestMessage(dataStructureUtil.randomBlobIdentifiers(5));
+        new BlobSidecarsByRootRequestMessage(schema, dataStructureUtil.randomBlobIdentifiers(5));
 
     handler.onIncomingMessage(protocolId, peer, request, callback);
 
@@ -173,7 +176,7 @@ public class BlobSidecarsByRootMessageHandlerTest {
         .thenReturn(SafeFuture.completedFuture(Optional.empty()));
 
     handler.onIncomingMessage(
-        protocolId, peer, new BlobSidecarsByRootRequestMessage(blobIdentifiers), callback);
+        protocolId, peer, new BlobSidecarsByRootRequestMessage(schema, blobIdentifiers), callback);
 
     verify(combinedChainDataClient, times(1)).getBlockByBlockRoot(secondBlockRoot);
     verify(callback, times(3)).respond(blobSidecarCaptor.capture());
@@ -211,7 +214,7 @@ public class BlobSidecarsByRootMessageHandlerTest {
                 Optional.of(dataStructureUtil.randomSignedBeaconBlock(UInt64.ONE))));
 
     handler.onIncomingMessage(
-        protocolId, peer, new BlobSidecarsByRootRequestMessage(blobIdentifiers), callback);
+        protocolId, peer, new BlobSidecarsByRootRequestMessage(schema, blobIdentifiers), callback);
 
     verify(callback, never()).respond(any());
     verify(callback).completeWithErrorResponse(rpcExceptionCaptor.capture());
@@ -239,7 +242,7 @@ public class BlobSidecarsByRootMessageHandlerTest {
         .thenReturn(SafeFuture.completedFuture(Optional.of(blobSidecar)));
 
     handler.onIncomingMessage(
-        protocolId, peer, new BlobSidecarsByRootRequestMessage(blobIdentifiers), callback);
+        protocolId, peer, new BlobSidecarsByRootRequestMessage(schema, blobIdentifiers), callback);
 
     verify(callback, never()).respond(any());
     verify(callback).completeWithErrorResponse(rpcExceptionCaptor.capture());
@@ -258,7 +261,7 @@ public class BlobSidecarsByRootMessageHandlerTest {
     final List<BlobIdentifier> blobIdentifiers = dataStructureUtil.randomBlobIdentifiers(5);
 
     handler.onIncomingMessage(
-        protocolId, peer, new BlobSidecarsByRootRequestMessage(blobIdentifiers), callback);
+        protocolId, peer, new BlobSidecarsByRootRequestMessage(schema, blobIdentifiers), callback);
 
     verify(callback, times(5)).respond(blobSidecarCaptor.capture());
